test task done Это объединение 5 коммитов.
added exception handling with ControllerAdvice

added security simple auth

test task done

deleted duild files

updated gitignore

+++ service/src/main/java/org/valerochka1337/service/RequestServiceImpl.java
@@ -1,12 +1,17 @@
 package org.valerochka1337.service;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.valerochka1337.entity.Request;
 import org.valerochka1337.entity.Status;
 import org.valerochka1337.entity.User;
+import org.valerochka1337.exceptions.request.EditNotDraftRequestException;
 import org.valerochka1337.exceptions.request.InvalidStatusRequestException;
 import org.valerochka1337.exceptions.request.NoSuchRequestException;
 import org.valerochka1337.exceptions.request.NotNULLIdCreationRequestException;
@@ -45,6 +50,9 @@ public class RequestServiceImpl implements RequestService {
       throw new NotNULLIdCreationRequestException();
     }
     Request requestEntity = requestMapper.toEntity(request);
+
+    requestEntity.setAuthor(getAuthorizedUser());
+    requestEntity.setStatus(Status.DRAFT);
     requestEntity.setCreationDate(LocalDateTime.now());
 
     requestEntity = requestRepository.save(requestEntity);
@@ -58,26 +66,61 @@ public class RequestServiceImpl implements RequestService {
       throw new NoSuchRequestException(id);
     }
 
+    if (!getAuthorizedUser()
+        .getId()
+        .equals(requestRepository.findById(id).get().getAuthor().getId())) {
+      throw new AccessDeniedException("Can't delete not owned request");
+    }
+
     requestRepository.deleteById(id);
   }
 
   @Override
-  public List<RequestModel> getRequests() {
-    return requestRepository.findAll().stream()
-        .map(requestMapper::toModel)
-        .collect(Collectors.toList());
+  public List<RequestModel> getRequests(Status status, Pageable pageable) {
+    User user = getAuthorizedUser();
+    List<Request> resultRequests = new ArrayList<>();
+    if (status == null) {
+      if (hasPermission(user, "requests:read:owned")) {
+        resultRequests = requestRepository.findAllByAuthor_Id(user.getId(), pageable);
+      }
+    } else {
+      if (status == Status.SENT && hasPermission(user, "requests:read:sent:all")) {
+        resultRequests = requestRepository.findAllByStatus(Status.SENT, pageable);
+      } else {
+        resultRequests =
+            requestRepository.findAllByStatusAndAuthor_Id(status, user.getId(), pageable);
+      }
+    }
+
+    if (user.getRoles().stream().anyMatch(role -> role.getName().compareTo("OPERATOR") == 0)) {
+      resultRequests =
+          resultRequests.stream()
+              .peek(req -> req.setMessage(messageFormatter.format(req.getMessage())))
+              .collect(Collectors.toList());
+    }
+
+    return resultRequests.stream().map(requestMapper::toModel).collect(Collectors.toList());
   }
 
   @Override
   public RequestModel getRequestById(Long id) {
     Request requestEntity =
         requestRepository.findById(id).orElseThrow(() -> new NoSuchRequestException(id));
+    User user = getAuthorizedUser();
+    if (requestEntity.getAuthor().getId().compareTo(user.getId()) != 0) {
+      throw new AccessDeniedException("Can't get request of another user");
+    }
+
+    if (user.getRoles().stream().anyMatch(role -> role.getName().compareTo("OPERATOR") == 0)) {
+      requestEntity.setMessage(messageFormatter.format(requestEntity.getMessage()));
+    }
 
     return requestMapper.toModel(requestEntity);
   }
 
   @Override
-  public List<RequestModel> getSentRequestsForUserByPartialUsername(String partialUsername) {
+  public List<RequestModel> getSentRequestsForUserByPartialUsername(
+      String partialUsername, Pageable pageable) {
     List<User> foundUsers = userRepository.findAllByUsernameContaining(partialUsername);
 
     if (foundUsers.isEmpty()) {
@@ -88,15 +131,31 @@ public class RequestServiceImpl implements RequestService {
     }
 
     User foundUser = foundUsers.get(0);
-    return foundUser.getRequests().stream()
-        .map(requestMapper::toModel)
-        .collect(Collectors.toList());
+    List<Request> resultRequest =
+        requestRepository.findAllByStatusAndAuthor_Id(Status.SENT, foundUser.getId(), pageable);
+    if (getAuthorizedUser().getRoles().stream()
+        .anyMatch(role -> role.getName().compareTo("OPERATOR") == 0)) {
+      resultRequest =
+          resultRequest.stream()
+              .peek(req -> req.setMessage(messageFormatter.format(req.getMessage())))
+              .collect(Collectors.toList());
+    }
+
+    return resultRequest.stream().map(requestMapper::toModel).collect(Collectors.toList());
   }
 
   @Override
   public RequestModel editMessageRequestWithId(Long id, String editedMessage) {
     Request existingRequest =
         requestRepository.findById(id).orElseThrow(() -> new NoSuchRequestException(id));
+
+    if (getAuthorizedUser().getRequests().stream()
+        .allMatch(request -> request.getId().compareTo(id) != 0)) {
+      throw new AccessDeniedException("Cannot edit not owned request");
+    }
+    if (existingRequest.getStatus().compareTo(Status.DRAFT) != 0) {
+      throw new EditNotDraftRequestException();
+    }
 
     existingRequest.setMessage(editedMessage);
     requestRepository.save(existingRequest);
@@ -108,6 +167,11 @@ public class RequestServiceImpl implements RequestService {
   public RequestModel sendRequestWithId(Long id) {
     Request existingRequest =
         requestRepository.findById(id).orElseThrow(() -> new NoSuchRequestException(id));
+
+    if (getAuthorizedUser().getRequests().stream()
+        .allMatch(request -> request.getId().compareTo(id) != 0)) {
+      throw new AccessDeniedException("Can't send not owned request");
+    }
 
     if (existingRequest.getStatus().compareTo(Status.DRAFT) != 0) {
       throw new InvalidStatusRequestException();
@@ -150,8 +214,20 @@ public class RequestServiceImpl implements RequestService {
     return requestMapper.toModel(existingRequest);
   }
 
-  private String formatMessageIfOperator(String message) {
-    // TODO
-    return messageFormatter.format(message);
+  private User getAuthorizedUser() {
+    String username = SecurityContextHolder.getContext().getAuthentication().getName();
+    return userRepository.findByUsername(username).orElseThrow(NoUserFoundUserException::new);
+  }
+
+  private boolean hasPermission(User user, String permissionName) {
+    for (var role : user.getRoles()) {
+      for (var permission : role.getPermissions()) {
+        if (permission.getName().equals(permissionName)) {
+          return true;
+        }
+      }
+    }
+
+    return false;
   }
 }
